Scroll wheel works with carousel, cleaned up code, more value in db
Added the carousel to both movies and shows. The scroll wheel works for both of them. Cleaned up and simplified some of the code in the details and index views. Show and movie controllers now pull more shows/movies from the db. Issue in show index view, will post to issues.